+++ src/android/CameraLauncher.java
@@ -119,6 +119,10 @@ public class CameraLauncher extends CordovaPlugin implements MediaScannerConnect
     private static final int CROP_GALERY = 666;
 
     private static final String TIME_FORMAT = "yyyyMMdd_HHmmss";
+    
+    //we need literal values because we cannot simply do Manifest.permission.READ_MEDIA_IMAGES, because of the target sdk
+    private static final String READ_MEDIA_IMAGES = "android.permission.READ_MEDIA_IMAGES";
+    private static final String READ_MEDIA_VIDEO = "android.permission.READ_MEDIA_VIDEO";
 
     private int mQuality;                   // Compression quality hint (0-100: 0=low quality & high compression, 100=compress of max quality)
     private int targetWidth;                // desired width of the image
@@ -134,7 +138,7 @@ public class CameraLauncher extends CordovaPlugin implements MediaScannerConnect
     private boolean orientationCorrected;   // Has the picture's orientation been corrected
     private boolean allowEdit;              // Should we allow the user to crop the image.
 
-    protected final static String[] permissions = { Manifest.permission.CAMERA, Manifest.permission.READ_EXTERNAL_STORAGE, Manifest.permission.WRITE_EXTERNAL_STORAGE };
+    protected final static String[] permissions = createPermissionArray();
 
     public CallbackContext callbackContext;
     private int numPics;
@@ -216,9 +220,14 @@ public class CameraLauncher extends CordovaPlugin implements MediaScannerConnect
                 }
                 else if ((this.srcType == PHOTOLIBRARY) || (this.srcType == SAVEDPHOTOALBUM)) {
                     // FIXME: Stop always requesting the permission
-                    if(!PermissionHelper.hasPermission(this, Manifest.permission.READ_EXTERNAL_STORAGE)) {
+                    if(Build.VERSION.SDK_INT < 33 && !PermissionHelper.hasPermission(this, Manifest.permission.READ_EXTERNAL_STORAGE)) {
                         PermissionHelper.requestPermission(this, SAVE_TO_ALBUM_SEC, Manifest.permission.READ_EXTERNAL_STORAGE);
-                    } else {
+                    }
+
+                    else if(Build.VERSION.SDK_INT >= 33 && (!PermissionHelper.hasPermission(this, READ_MEDIA_IMAGES) || !PermissionHelper.hasPermission(this, READ_MEDIA_VIDEO))){
+                        PermissionHelper.requestPermissions(this, SAVE_TO_ALBUM_SEC, new String[]{ READ_MEDIA_VIDEO, READ_MEDIA_IMAGES});
+                    }
+                    else {
                         this.getImage(this.srcType, destType, encodingType);
                     }
                 }
@@ -266,8 +275,14 @@ public class CameraLauncher extends CordovaPlugin implements MediaScannerConnect
      * @param encodingType           Compression quality hint (0-100: 0=low quality & high compression, 100=compress of max quality)
      */
     public void callTakePicture(int returnType, int encodingType) {
-        boolean saveAlbumPermission = PermissionHelper.hasPermission(this, Manifest.permission.READ_EXTERNAL_STORAGE)
-                && PermissionHelper.hasPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE);
+
+        boolean saveAlbumPermission = (Build.VERSION.SDK_INT < 33 &&
+                PermissionHelper.hasPermission(this, Manifest.permission.READ_EXTERNAL_STORAGE) &&
+                PermissionHelper.hasPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE)) ||
+                (Build.VERSION.SDK_INT >= 33 &&
+                        PermissionHelper.hasPermission(this, READ_MEDIA_VIDEO) &&
+                        PermissionHelper.hasPermission(this, READ_MEDIA_IMAGES));
+
         boolean takePicturePermission = PermissionHelper.hasPermission(this, Manifest.permission.CAMERA);
 
         // CB-10120: The CAMERA permission does not need to be requested unless it is declared
@@ -297,10 +312,14 @@ public class CameraLauncher extends CordovaPlugin implements MediaScannerConnect
             takePicture(returnType, encodingType);
         } else if (saveAlbumPermission && !takePicturePermission) {
             PermissionHelper.requestPermission(this, TAKE_PIC_SEC, Manifest.permission.CAMERA);
-        } else if (!saveAlbumPermission && takePicturePermission) {
+        } else if (!saveAlbumPermission && takePicturePermission && Build.VERSION.SDK_INT < 33) {
             PermissionHelper.requestPermissions(this, TAKE_PIC_SEC,
                     new String[]{Manifest.permission.READ_EXTERNAL_STORAGE, Manifest.permission.WRITE_EXTERNAL_STORAGE});
-        } else {
+        } else if(!saveAlbumPermission && takePicturePermission && Build.VERSION.SDK_INT >= 33){
+            PermissionHelper.requestPermissions(this, TAKE_PIC_SEC,
+                    new String[]{READ_MEDIA_VIDEO, READ_MEDIA_IMAGES});
+        }
+        else {
             PermissionHelper.requestPermissions(this, TAKE_PIC_SEC, permissions);
         }
     }
@@ -1460,8 +1479,15 @@ public class CameraLauncher extends CordovaPlugin implements MediaScannerConnect
                     cropIntent,
                     requestCode + destType);
         }
+    }
 
-
+    private static String[] createPermissionArray(){
+        if(Build.VERSION.SDK_INT < 33){
+            return new String[]{Manifest.permission.CAMERA, Manifest.permission.READ_EXTERNAL_STORAGE, Manifest.permission.WRITE_EXTERNAL_STORAGE};
+        }
+        else{
+            return new String[]{Manifest.permission.CAMERA, READ_MEDIA_IMAGES, READ_MEDIA_VIDEO};
+        }
     }
 
 }