+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 The changes documented here do not include those from the original repository.
 
+## [Unreleased]
+### Fixes
+- Added permission requests for Android 13. (https://outsystemsrd.atlassian.net/browse/RMET-1831)
+
 ## [4.2.0-OS39]
 ### Fixes
 - Update error codes and messages for iOS. (https://outsystemsrd.atlassian.net/browse/RMET-1744)